centered items via flexbox
I've taken another aproach into centering the items on the screen. I see that you've used margins but the most common and practical way to do so is using display: flex. You'll see it next in the course so don't worry! anyways this is not the absolute best way to do centering, take it as an example...